Support to upload image for product, view a product, add exeception handler

+++ src/main/java/mum/waa/coffee/domain/OrderLine.java
@@ -51,11 +51,11 @@ public class OrderLine {
     }
 
     public double getSubtotal() {
-        return quantity * product.getPrice();
+        return quantity * product.getPrice().doubleValue();
     }
 
     public double getPrice() {
-        return product.getPrice();
+        return product.getPrice().doubleValue();
     }
 
     @Override
